My Progress
This is what I have done so far. The main view is just something I was
working on the other day, I will probably replace it with yours once you
upload it. You can click on the more button though to see the stats view
that I am working on. I created a Stat object that stores a name and a
value so that stats can be added to the list view using the StatAdapter.