Created the goal adapter and the xml to go with it. Have not implemented them yet.
Also added an attribute to the Goals object to store the units that a goal was entered in to provide for better goal saving. The default unit is miles.
I created another toString method as well to format the goals so that they can be displayed in the list of suggested goals.